% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/CheckData.R
\name{CheckData}
\alias{CheckData}
\title{Check the form and basic structure of the functional data 'y' and the recorded times 'tt'.}
\usage{
CheckData(y, tt)
}
\arguments{
\item{y}{is a n-by-1 list of vectors}

\item{tt}{is a n-by-1 list of vectors}
}
\value{
logical
}
\description{
Check the form and basic structure of the functional data 'y' and the recorded times 'tt'.
}
\examples{
1 + 3
}

