% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/CheckOptions.R
\name{CheckOptions}
\alias{CheckOptions}
\title{Check if the options structure is valid and set the ones that are NULL}
\usage{
CheckOptions(t, optns, n)
}
\arguments{
\item{n}{is a total number of sample curves}

\item{p}{is an initialized option list}
}
\value{
logical
}
\description{
Check if the options structure is valid and set the ones that are NULL
}
\examples{
1 + 3
}

