% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/createBetaPlots.R
\name{createBetaPlots}
\alias{createBetaPlots}
\title{Create estimates for the beta functions used}
\usage{
createBetaPlots(fpcaRegObj, sList, openNewDev = TRUE)
}
\arguments{
\item{fpcaRegObj}{A object of class FPCAreg returned by the function FPCAregScalar().}

\item{sList}{A list with the support of each function / If vector then the same support is used for all betas.}
}
\description{
Create estimates for the beta functions used in a FPCA regression object
}

