% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/createCovPlot.R
\name{createCovPlot}
\alias{createCovPlot}
\title{Create the covariance surface plot based on the results from FPCA() or FPCder().}
\usage{
createCovPlot(yy, corrPlotType = "Fitted", isInteractive = FALSE)
}
\arguments{
\item{yy}{returned object from FPCA().}

\item{corrPlotType}{a string specifying the type of covariance surface to be plotted:
'Smoothed': plot the smoothed cov surface
'Fitted': plot the fitted cov surface
'Raw': plot the raw covariance scatter plot}

\item{isInteractive}{an option for interactive plot:
TRUE: interactive plot; FALSE: printable plot}

\item{...}{other arguments passed into persp3d, persp3D, plot3d or points3D for plotting options}
}
\description{
This function will open a new device if not instructed otherwise.
}
\examples{
set.seed(1)
n <- 20
pts <- seq(0, 1, by=0.05)
sampWiener <- wiener(n, pts)
sampWiener <- sparsify(sampWiener, pts, 10)
res <- FPCA(sampWiener$yList, sampWiener$tList, list(dataType='Sparse', error=FALSE, kernel='epan', verbose=TRUE))
createCovPlot(res)
}

