% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/createFuncBoxPlot.R
\name{createFuncBoxPlot}
\alias{createFuncBoxPlot}
\title{Create functional boxplot using 'bagplot' or 'pointwise' methodology}
\usage{
createFuncBoxPlot(fpcaObj, addIndx = NULL, variant = "bagplot", ...)
}
\arguments{
\item{fpcaObj}{A object of class FPCA returned by the function FPCA().}

\item{addIndx}{A vector of indeces corresponding to which samples one should overlay (Default: NULL)}

\item{variant}{A character variable indicating which methodology should be used ('bagplot' or 'pointwise')to create the functional box-plot (Default: 'bagplot')}

\item{...}{Additional arguments for the 'plot' function.}
}
\description{
Using an FPCA object create a functional box-plot based on the function scores.
}
\examples{
set.seed(1)
n <- 20
pts <- seq(0, 1, by=0.05)
sampWiener <- wiener(n, pts)
sampWiener <- sparsify(sampWiener, pts, 10)
res <- FPCA(sampWiener$yList, sampWiener$tList,
            list(dataType='Sparse', error=FALSE, kernel='epan', verbose=TRUE))
createFuncBoxPlot(res, addIndx=c(1:3))
}
\references{
\cite{P. J. Rousseeuw, I. Ruts, J. W. Tukey (1999): The bagplot: a bivariate boxplot, The American Statistician, vol. 53, no. 4, 382-387}
}

