% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createModeOfVarPlot.R
\name{createModeOfVarPlot}
\alias{createModeOfVarPlot}
\title{Functional Principal Component Analysis mode of variation plot}
\usage{
createModeOfVarPlot(ret, k = 1, ...)
}
\arguments{
\item{ret}{An FPCA class object returned by FPCA().}

\item{k}{The k-th mode of variation to plot (default k = 1)}
}
\description{
This function will open a new device if not instructed otherwise.
}
\examples{
set.seed(1)
n <- 20
pts <- seq(0, 1, by=0.05)
sampWiener <- wiener(n, pts)
sampWiener <- sparsify(sampWiener, pts, 10)
res <- FPCA(sampWiener$yList, sampWiener$tList, list(dataType='Sparse', error=FALSE, kernel='epan', verbose=TRUE))
createDiagnosticsPlot(res)
}

