% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/CreateOptions.R
\name{CreateOptions}
\alias{CreateOptions}
\title{Create the options list used by FPCA}
\usage{
CreateOptions(bwmu = NULL, bwmuMethod = NULL, bwuserCov = NULL,
  bwuserCovGcv = NULL, ntest1 = NULL, selectionMethod = NULL,
  FVEthreshold = NULL, numComponents = NULL, maxK = NULL,
  dataType = NULL, error = NULL, nRegGrid = NULL, methodXi = NULL,
  shrink = NULL, kernel = NULL, numBins = NULL, yname = NULL,
  diagnosticsPlot = NULL, rho = NULL, verbose = NULL, userMu = NULL,
  userCov = NULL, methodCov = NULL, methodMu = NULL, outPercent = NULL,
  useBinnedData = NULL, rotationCut = NULL)
}
\arguments{
\item{bwmu}{: bandwidth value for mean function is using CV or GCV; positive numeric - default: determine automatically based on 'bwmuMethod'}

\item{bwmuMethod}{: bandwidth choice method for mean function; 'GMeanAndGCV','CV','GCV - default: 'GMeanAndGCV''}

\item{ntest1}{: number of curves used for CV when choosing bandwidth; [1,N] - default: min(30, N-1), N : # of curves}

\item{selectionMethod}{: the method of choosing the number of principal components K; 'FVE','AIC','BIC' : default 'FVE' - only 'FVE' avaiable now/ default 'FVE')}

\item{FVEthreshold}{: Fraction-of-Variance-Explained threshold used during the SVD of the fitted covar. function; numeric (0,1] - default: 0.9999}

\item{maxK}{: maximum number of principal components to consider; positive integer - default: min(20, N-1), N : # of curves}

\item{dataType}{: do we have sparse or dense functional data; 'Sparse', 'Dense', 'DenseWithMV', 'p>>n' - default:  determine automatically based on 'IsRegular'}

\item{error}{: assume measurement error in the dataset; logical - default: TRUE}

\item{nRegGrid}{: number of support points in each direction of covariance surface; numeric - default: 51}

\item{methodXi}{: method to estimate the PC scores; 'CE', 'IN' - default: 'CE'}

\item{shrink}{: apply shrinkage to estimates of random coefficients (dense data only); logical - default: FALSE}

\item{kernel}{: smoothing kernel choice, common for mu and covariance; "rect", "gauss", "epan", "gausvar", "quar" - default: "epan" for dense data else "gauss"}

\item{numBins}{: number of bins to bin the data into; positive integer > 10, default: NULL}

\item{rho}{: truncation threshold for the iterative residual. 'cv': choose rho by leave-one-observation out cross-validation; 'no': use the iterative sigma2 estimate - default "cv".}

\item{verbose}{: display diagnostic messages; logical - default: FALSE}

\item{userMu}{: user-defined smoothed mean function; numerical vector - default: NULL}

\item{userCov}{: user-defined smoothed covariance function; numerical matrix - default: NULL}

\item{methodCov}{:  method to estimate covariance; 'PACE','RARE','CrossSectional' - automatically determined, user input ignored}

\item{methodMu}{:  method to estimate mu; 'PACE','RARE','CrossSectional' - automatically determined, user input ignored}

\item{outPercent}{: 2-element vector in [0,1] indicating the outPercent data in the boundary - default (0,1)}

\item{useBinnedData}{: 'FORCE' (Enforce the # of bins), 'AUTO' (Select the # of  bins automatically), 'OFF' (Do not bin) - default: 'AUTO'}

\item{rotationCut}{: 2-element vector in [0,1] indicating the percent of data truncated during sigma^2 estimation; default  (0.25, 0.75))}

\item{bwcov}{: bandwidth value for covariance function; positive numeric - default: determine automatically based on 'bwcovMethod'}

\item{bwcovMethod}{: bandwidth choice method for covariance function; 'GMeanAndGCV','CV','GCV - default: 'GMeanAndGCV'')}

\item{useBins:}{testing purpose: whether to bin the same observed time points when 2D smoothing; logical - default: FALSE}
}
\value{
an option list
}
\description{
Create the options list used by FPCA
}
\examples{
optLst = CreateOptions(kernel='rect');  # Create options list with rectangular kernel
}

