% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createOutliersPlot.R
\name{createOutliersPlot}
\alias{createOutliersPlot}
\title{Functional Principal Component Scores Plot}
\usage{
createOutliersPlot(fpcaObj, factor = 3, ...)
}
\arguments{
\item{fpcaObj}{An FPCA class object returned by FPCA().}

\item{factor}{Inflation factor defining the loop (see ?compute.bagplot).}
}
\description{
This function will create using the first two FPC scores a set of convex hulls of the scores as these hulls are defined by the references.
}
\examples{
set.seed(1)
n <- 20
pts <- seq(0, 1, by=0.05)
sampWiener <- wiener(n, pts)
sampWiener <- sparsify(sampWiener, pts, 10)
res <- FPCA(sampWiener$yList, sampWiener$tList, list(dataType='Sparse', error=FALSE, kernel='epan', verbose=TRUE))
createOutliersPlot(res)
}
\references{
\cite{P. J. Rousseeuw, I. Ruts, J. W. Tukey (1999): The bagplot: a bivariate boxplot, The American Statistician, vol. 53, no. 4, 382-387}
}

