% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createPathPlot.R
\name{createPathPlot}
\alias{createPathPlot}
\title{Create the covariance surface plot based on the results from FPCA() or FPCder().}
\usage{
createPathPlot(fpcaObj, subset, k = NULL, inputData = fpcaObj$inputData,
  showObs = !is.null(inputData), ...)
}
\arguments{
\item{fpcaObj}{returned object from FPCA().}

\item{inputData}{A list of length 2 containing the sparse/dense (unsupported yet) observations. \code{inputData} needs to contain two fields: \code{t} for a list of time points and \code{y} for a list of observations.}

\item{...}{other arguments passed into matplot for plotting options}
}
\description{
This function will open a new device if not instructed otherwise.
}
\examples{
set.seed(1)
n <- 20
pts <- seq(0, 1, by=0.05)
sampWiener <- wiener(n, pts)
sampWiener <- sparsify(sampWiener, pts, 10)
res <- FPCA(sampWiener$yList, sampWiener$tList, 
            list(dataType='Sparse', error=FALSE, kernel='epan', verbose=TRUE))
createPathPlot(res, subset=1:5)
}

