% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DynCorr.R
\name{DynCorr}
\alias{DynCorr}
\title{Dynamical Correlation}
\usage{
DynCorr(x, y, t)
}
\arguments{
\item{x}{a n by m matrix where rows representing subjects and columns representing measurements, missings are allowed.}

\item{y}{a n by m matrix where rows representing subjects and columns representing measurements, missings are allowed.}

\item{t}{a length m vector of time points where x,y are observed.}
}
\value{
A length m vector of individual dynamic correlations
}
\description{
Calculate Dynamical Correlation for 2 paired dense regular functional data observed on the same grid.
}
\examples{
library(MASS)
n=200             # sample size
t=seq(0,1,length.out=100)       # length of data
mu_quad_x=8*t^2-4*t+5
mu_quad_y=8*t^2-12*t+6
fun=rbind(rep(1,length(t)),-t,t^2)
z1=mvrnorm(n,rep(0,3),diag(c(2,16/3,4)))
x1_quad_error=y1_quad_error=matrix(0,nrow=n,ncol=length(t))
for (i in 1:n){
 x1_quad_error[i,]=mu_quad_x+z1[i,]\%*\%fun+rnorm(length(t),0,0.01)
 y1_quad_error[i,]=mu_quad_y+2*z1[i,]\%*\%fun +rnorm(length(t),0,0.01)
}
dyn1_quad=DynCorr(x1_quad_error,y1_quad_error,t) 

}
\references{
\cite{Dubin J A, M\"uller H G. Dynamical correlation for multivariate longitudinal data[J]. Journal of the American Statistical Association, 2005, 100(471): 872-881.}

\cite{Liu S, Zhou Y, Palumbo R, et al. Dynamical correlation: A new method for quantifying synchrony with multivariate intensive longitudinal data[J]. Psychological methods, 2016, 21(3): 291.}
}
