% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitted.FPCA.R
\name{fitted.FPCA}
\alias{fitted.FPCA}
\title{Fitted functional sample from FPCA object}
\usage{
\method{fitted}{FPCA}(object, k = NULL, der = FALSE, method = NULL, ...)
}
\arguments{
\item{object}{A object of class FPCA returned by the function FPCA().}

\item{k}{The integer number of the first k components used for the representation. (default: length(fpcaObj$lambda ))}

\item{der}{A logical specifying if derivatives should be returned or instead of the fitted values (default: FALSE)}

\item{method}{The method used to produce the sample of derivatives ('EIG' (default) or 'QUO'). See Liu and Mueller (2009) for more details}

\item{...}{Additional arguments}
}
\description{
Combine the zero-meaned fitted values and the interpolated mean to get the fitted values for the trajectories or the first derivatives of these trajectories.
}
\examples{
set.seed(1)
n <- 20
pts <- seq(0, 1, by=0.05)
sampWiener <- wiener(n, pts)
sampWiener <- sparsify(sampWiener, pts, 10)
res <- FPCA(sampWiener$yList, sampWiener$tList, 
            list(dataType='Sparse', error=FALSE, kernel='epan', verbose=TRUE))
fittedY <- fitted(res)
}
\references{
\cite{Liu, Bitao, and Hans-Georg Mueller. "Estimating derivatives for samples of sparsely observed functions, with application to online auction dynamics." Journal of the American Statistical Association 104, no. 486 (2009): 704-717. (Sparse data FPCA)}
}

