% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fitted.FRegObj.R
\name{fitted.FregObj}
\alias{fitted.FregObj}
\title{Fitted functional sample from FPCA regression object}
\usage{
\method{fitted}{FregObj}(object, fpcaObj, scaleZ = FALSE, expVarScal,
  expVarFunc, ...)
}
\arguments{
\item{fpcaObj}{A object of class FPCA returned by the function FPCA().}

\item{fpcaObjder}{A object of class FPCAder returned by the function FPCAder(.}
}
\description{
Combine the zero-meaned fitted values and the interpolated mean to get the final values.
}
\examples{
set.seed(1)
n <- 20
pts <- seq(0, 1, by=0.05)
sampWiener <- wiener(n, pts)
sampWiener <- sparsify(sampWiener, pts, 10)
res <- FPCA(sampWiener$yList, sampWiener$tList,
            list(dataType='Sparse', error=FALSE, kernel='epan', verbose=TRUE))
fittedY <- fitted(res)
}
\references{
\cite{Liu, Bitao, and Hans-Georg Mueller. "Estimating derivatives for samples of sparsely observed functions, with application to online auction dynamics." Journal of the American Statistical Association 104, no. 486 (2009): 704-717. (Sparse data FPCA)}
}

