% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FPCAregFuncExp.R
\name{FPCAregFuncExp}
\alias{FPCAregFuncExp}
\title{Functional Principal Component Analysis Regression with Functional dependent variable}
\usage{
FPCAregFuncExp(depVar, expVarScal = NULL, expVarFunc = NULL,
  regressionType = NULL, bwScalar = NULL, bwFunct = NULL,
  fastSmooth = FALSE, verbose = FALSE)
}
\arguments{
\item{depVar}{An FPCA object}

\item{expVarScal}{A data.frame holding scalar explanatory variables; NAs will be omitted internally.}

\item{expVarFunc}{A list holding the FPCA objects for each of the functional explanatory variables.}

\item{regressionType}{A string defining the type of regression to perform ('dense' or 'sparse'); (default : automatically determined based on 'depVar')}

\item{bwScalar}{The value of bandwidth to be used for all scalar/functional cross-covariances (default: automatically determined using GCV)}

\item{bwFunct}{The values of bandwiths to be used for all function/function cross-covariances (default: automatically determined using GCV)}

\item{fastSmooth}{Use thin-plate splines during the estimation of the cross-covariance (default: FALSE)}

\item{verbose}{If TRUE print out the bandwidth used during the GCV procedures of selecting them}

\item{...}{Additional arguments}
}
\description{
Functional regression for dense or sparse functional data.
}
\references{
\cite{Yao, F., Mueller, H.G., Wang, J.L. "Functional Linear Regression Analysis for Longitudinal Data." Annals of Statistics 33, (2005): 2873-2903.(Dense data)}
\cite{Senturk, D., Nguyen, D.V. "Varying Coefficient Models for Sparse Noise-contaminated Longitudinal Data", Statistica Sinica 21(4), (2011): 1831-1856. (Sparse data)}
}

