% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FPCAregScalar.R
\name{FPCAregScalar}
\alias{FPCAregScalar}
\title{Functional Principal Component Analysis Regression with Scalar dependent variable}
\usage{
FPCAregScalar(fpcaObj, extVar = NULL, depVar, varSelect = NULL,
  crossVal = FALSE, regressionType = NULL, ...)
}
\arguments{
\item{fpcaObj}{A object of class FPCA returned by the function FPCA().}

\item{extVar}{A data.frame holding external explanatory variables.}

\item{depVar}{A vector with the dependant variable.}

\item{varSelect}{A string defining the type of step-wise variable selection applied ('AIC' or 'BIC'); this calls 'MASS::stepAIC()'. (default: NULL)}

\item{crossVal}{A logical variable indicating if cross-validation measures for the predictive accuracy should be returned; this calls 'caret::train()'. (default: FALSE)}

\item{regressionType}{A string defining the type of regression to perform ('dense' or 'sparse'); (default : automatically determined based on 'fpcaObj'}

\item{...}{Additional arguments}
}
\description{
Functional regression for dense or sparse functional data.
}
\details{
If both 'varSelect' and 'crossVal' are used then the cross-validation is applied to the model returned by the variable selection procedure.
}
\references{
\cite{Yao, F., Mueller, H.G., Wang, J.L. "Functional Linear Regression Analysis for Longitudinal Data." Annals of Statistics 33, (2005): 2873-2903.(Dense data)}
}

