% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FSVD.R
\name{FSVD}
\alias{FSVD}
\title{Functional Singular Value Decomposition}
\usage{
FSVD(Ly1, Lt1, Ly2, Lt2, FPCAoptns1 = NULL, FPCAoptns2 = NULL,
  SVDoptns = list())
}
\arguments{
\item{Ly1}{A list of \emph{n} vectors containing the observed values for each individual. Missing values specified by \code{NA}s are supported for dense case (\code{dataType='dense'}).}

\item{Lt1}{A list of \emph{n} vectors containing the observation time points for each individual corresponding to y. Each vector should be sorted in ascending order.}

\item{Ly2}{A list of \emph{n} vectors containing the observed values for each individual. Missing values specified by \code{NA}s are supported for dense case (\code{dataType='dense'}).}

\item{Lt2}{A list of \emph{n} vectors containing the observation time points for each individual corresponding to y. Each vector should be sorted in ascending order.}

\item{FPCAoptns1}{A list of options control parameters specified by \code{list(name=value)} for the FPC analysis of sample 1. See `?FPCA'.}

\item{FPCAoptns2}{A list of options control parameters specified by \code{list(name=value)} for the FPC analysis of sample 2. See `?FPCA'.}

\item{SVDoptns}{A list of options control parameters specified by \code{list(name=value)} for the FSVD analysis of samples 1 & 2. See `Details`.}
}
\value{
A list containing the following fields:
\item{bw1}{The selected (or user specified) bandwidth for smoothing the cross-covariance function across the support of sample 1.}
\item{bw2}{The selected (or user specified) bandwidth for smoothing the cross-covariance function across the support of sample 2.}
\item{CrCov}{The smoothed cross-covariance between samples 1 & 2.}
\item{sValues}{A list of length \emph{nsvd}, each entry containing the singuar value estimates for the FSC estimates.}
\item{nsvd}{The number of singular componentes used.}
\item{canCorr}{The canonical correlations for each dimension.}
\item{FVE}{A percentage indicating the total variance explained by chosen FSCs with corresponding 'FVEthreshold'.}
\item{sFun1}{An nWorkGrid by \emph{K} matrix containing the estimated singular functions for sample 1.}
\item{sFun2}{An nWorkGrid by \emph{K} matrix containing the estimated singular functions for sample 2.}
\item{grid1}{A vector of length nWorkGrid1. The internal regular grid on which the singular analysis is carried on the support of sample 1.}
\item{grid2}{A vector of length nWorkGrid2. The internal regular grid on which the singular analysis is carried on the support of sample 2.}
\item{sScores1}{A \emph{n} by \emph{K} matrix containing the singular scores for sample 1.}
\item{sScores2}{A \emph{n} by \emph{K} matrix containing the singular scores for sample 2.}
\item{optns}{A list of options used by the SVD and the FPCA's procedures.}
}
\description{
FSVD for a pair of dense or sparse functional data.
}
\details{
Available control options for SVDoptns are: 
\describe{
\item{userBwCov}{The bandwidth value for the smoothed covariance function; positive numeric - default: determine automatically based on 'methodBwCov'}
\item{methodBwCov}{The bandwidth choice method for the smoothed covariance function; 'GMeanAndGCV' (the geometric mean of the GCV bandwidth and the minimum bandwidth),'CV','GCV' - default: 10\% of the support}
\item{userBwMu}{The bandwidth value for the smoothed mean function (using 'CV' or 'GCV'); positive numeric - default: determine automatically based on 'methodBwMu'}
\item{blahblah}{More to follow.}
\item{noScores}{Logical describing...}
}
}

