% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetCrCorYX.R
\name{GetCrCorYX}
\alias{GetCrCorYX}
\title{Make cross-correlation matrix from cross-covariance matrix}
\usage{
GetCrCorYX(ccXY, ccXX, ccYY)
}
\arguments{
\item{ccXY}{The cross-covariance matrix between variables X and Y.}

\item{ccXX}{The auto-covariance matrix of variable X or the diagonal of that matrix.}

\item{ccYY}{The auto-covariance matrix of variable Y or the diagonal of that matrix.}
}
\description{
Make cross-correlation matrix from cross-covariance matrix
}

