% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetCrCorYZ.R
\name{GetCrCorYZ}
\alias{GetCrCorYZ}
\title{Make cross-correlation matrix from auto- and cross-covariance matrix}
\usage{
GetCrCorYZ(ccYZ, acYY, covZ)
}
\arguments{
\item{ccYZ}{The cross-covariance vector between variables Y and Z (n-by-1).}

\item{acYY}{The auto-covariance n-by-n matrix of variable Y or the (n-by-1) diagonal of that matrix.}

\item{covZ}{The (scalar) covariance of variable Z.}
}
\value{
A cross-correlation matrix between variables Y (functional) and Z (scalar).
}
\description{
Make cross-correlation matrix from auto- andcross-covariance matrix
}
