% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetLWSigma2.R
\name{GetLWSigma2}
\alias{GetLWSigma2}
\title{Estimation of measurement error \code{sigma2} via Z Lin and JL Wang (2021+)}
\usage{
GetLWSigma2(y, t)
}
\arguments{
\item{y}{A list of \emph{n} vectors containing the observed values for each individual.}

\item{t}{A list of \emph{n} vectors containing the observation time points for each individual corresponding to y.}
}
\value{
\item{sigma2}{Variance for measurement error.}
}
\description{
Estimation of measurement error \code{sigma2} via Z Lin and JL Wang (2021+)
}
\references{
\cite{Z Lin and JL Wang (2021+). "Mean and covariance estimation for functional snippets." JASA.}
}
