% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/HandleNumericsAndNAN.R
\name{HandleNumericsAndNAN}
\alias{HandleNumericsAndNAN}
\title{#Check if NaN are present in the data and if yes remove them
#
#Check if there are problems cause by missing values with the form and basic structure of the functional data 'Ly' and the recorded times 'Lt'.
#
#}
\usage{
HandleNumericsAndNAN(Ly, Lt)
}
\arguments{
\item{Ly}{is a n-by-1 list of vectors
#}

\item{Lt}{is a n-by-1 list of vectors}
}
\description{
#Check if NaN are present in the data and if yes remove them
#
#Check if there are problems cause by missing values with the form and basic structure of the functional data 'Ly' and the recorded times 'Lt'.
#
#
}

