% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lwls1D.R
\name{Lwls1D}
\alias{Lwls1D}
\title{One dimensional local linear kernel smoother}
\usage{
Lwls1D(bw, kernel_type, win, xin, yin, xout, npoly = 1L, nder = 0L)
}
\arguments{
\item{bw}{Scalar holding the bandwidth}

\item{kernel_type}{Character holding the kernel type (see ?FPCA for supported kernels)}

\item{win}{Vector of length N with weights}

\item{xin}{Vector of length N with measurement points}

\item{yin}{Vector of length N with measurement values}

\item{xout}{Vector of length M with output measurement points}

\item{npoly}{Scalar (integer) degree of polynomial fitted (default 1)}

\item{nder}{Scalar (integer) degree of derivative fitted (default 0)}
}
\value{
Vector of length M with measurement values at the the point speficied by 'xout'
}
\description{
One dimensional local linear kernel smoother for longitudinal data.
}

