% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lwls2DDeriv.R
\name{Lwls2DDeriv}
\alias{Lwls2DDeriv}
\title{Two dimensional local linear kernel smoother to target derivatives.}
\usage{
Lwls2DDeriv(bw, kern = "epan", xin, yin, win = NULL, xout1 = NULL,
  xout2 = NULL, xout = NULL, npoly = 1L, nder1 = 0L, nder2 = 0L,
  subset = NULL, crosscov = TRUE, method = "sort2")
}
\arguments{
\item{bw}{A scalar or a vector of length 2 specifying the bandwidth.}

\item{kern}{Kernel used: 'gauss', 'rect', 'gausvar', 'epan' (default), 'quar'.}

\item{xin}{An n by 2 data frame or matrix of x-coordinate.}

\item{yin}{A vector of y-coordinate.}

\item{win}{A vector of weights on the observations.}

\item{xout1}{a p1-vector of first output coordinate grid. Defaults to the input gridpoints if left unspecified.}

\item{xout2}{a p2-vector of second output coordinate grid. Defaults to the input gridpoints if left unspecified.}

\item{xout}{alternative to xout1 and xout2. A matrix of p by 2 specifying the output points (may be inefficient if the size of \code{xout} is small).}

\item{npoly}{The degree of polynomials (include all \eqn{x^a y^b} terms where \eqn{a + b <= npoly})}

\item{nder1}{Order of derivative in the first direction}

\item{nder2}{Order of derivative in the second direction}

\item{subset}{a vector with the indices of x-/y-/w-in to be used (Default: NULL)}

\item{crosscov}{using function for cross-covariance estimation (Default: TRUE)}

\item{method}{should one try to sort the values xin and yin before using the lwls smoother? if yes ('sort2' - default for non-Gaussian kernels), if no ('plain' - fully stable; de)}
}
\value{
a p1 by p2 matrix of fitted values if xout is not specified. Otherwise a vector of length p corresponding to the rows of xout.
}
\description{
Two dimensional local weighted least squares smoother. Only a local linear smoother for estimating the original curve is available (no higher order)
}
