% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeBWtoZscore02y.R
\name{MakeBWtoZscore02y}
\alias{MakeBWtoZscore02y}
\title{Z-score body-weight for age 0 to 24 months based on WHO standards}
\usage{
MakeBWtoZscore02y(sex, age, bw)
}
\arguments{
\item{sex}{A character 'M' or 'F' indicating the sex of the child.}

\item{age}{A vector of time points of size Q.}

\item{bw}{A vector of body-weight readings of size Q.}
}
\value{
A vector of Z-scores of size Q.
}
\description{
Make vector of age and body-weight to z-scores based on WHO standards using LMS
}

