% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeHCtoZscore02y.R
\name{MakeHCtoZscore02y}
\alias{MakeHCtoZscore02y}
\title{Z-score head-circumference for age 0 to 24 months based on WHO standards}
\usage{
MakeHCtoZscore02y(sex, age, hc)
}
\arguments{
\item{sex}{A character 'M' or 'F' indicating the sex of the child.}

\item{age}{A vector of time points of size Q.}

\item{hc}{A vector of head circumference readings of size Q (in cm).}
}
\value{
A vector of Z-scores of size Q.
}
\description{
Make vector of age and height measurement to z-scores based on WHO standards using mu and sigma (not LMS)
}

