% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeLNtoZscore02y.R
\name{MakeLNtoZscore02y}
\alias{MakeLNtoZscore02y}
\title{Z-score height for age 0 to 24 months based on WHO standards}
\usage{
MakeLNtoZscore02y(sex, age, ln)
}
\arguments{
\item{sex}{A character 'M' or 'F' indicating the sex of the child.}

\item{age}{A vector of time points of size Q.}

\item{ln}{A vector of body-length readings of size Q (in cm).}
}
\value{
A vector of Z-scores of size Q.
}
\description{
Convert vector of age and height measurement to z-scores based on WHO standards using mu and sigma (not LMS)
}
