% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormCurveToArea.R
\name{NormCurvToArea}
\alias{NormCurvToArea}
\title{Normalize a curve to a particular area, by multiplication with a factor}
\usage{
NormCurvToArea(y, x = seq(0, 1, length.out = length(y)), area = 1)
}
\arguments{
\item{y}{values of curve at time-points \code{x}}

\item{x}{design time-points (default: \code{seq(0,1, length.out=length(y))})}

\item{area}{value to normalize the curve onto (default: 1)}
}
\value{
values of curve at times \code{x} such that its integration over \code{x} equals \code{area}.
}
\description{
Normalize a curve such that its integration over the design time-points equals a particular value (according to trapezoid integration).
}
