% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.FPCA.R
\name{plot.FPCA}
\alias{plot.FPCA}
\title{Plot an FPCA object.}
\usage{
\method{plot}{FPCA}(x, ...)
}
\arguments{
\item{x}{An FPCA class object returned by FPCA()}

\item{...}{passed into CreateDiagnosticsPlot}
}
\description{
Currently implemented as just plotting the diagnostics plots, the same as CreateDiagnosticsPlot.
}
\seealso{
CreateDiagnosticsPlot
}

