% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.FPCA.R
\name{predict.FPCA}
\alias{predict.FPCA}
\title{Predict FPC scores for a new sample given an FPCA object}
\usage{
\method{predict}{FPCA}(x, newX, newY, sigma2 = NULL, k = 1, im = "CE",
  ...)
}
\arguments{
\item{x}{An FPCA object.}

\item{newX}{A list of \emph{n} vectors containing the observation time points for each individual corresponding to y.}

\item{newY}{A list of \emph{n} vectors containing the observed values for each individual.}

\item{sigma2}{The user-defined measurement error variance. A positive scalar. (default: rho if applicable, otherwise sigma2 if applicable, otherwise 0 if no error. )}

\item{k}{The scalar defining the number of clusters to define; (default: 1).}

\item{im}{The integration method used to calculate the functional principal component scores ( standard numerical integration 'IN' or conditional expectation 'CE'); default: 'CE'.}

\item{...}{Not used.}
}
\value{
A matrix of size n-by-k
}
\description{
Return a matrix with the first k FPC scores according to conditional expectation or numerical integration.
}
\examples{
set.seed(1)
n <- 20
pts <- seq(0, 1, by=0.05)
sampWiener <- Wiener(n, pts)
sampWiener <- Sparsify(sampWiener, pts, 10)
res <- FPCA(sampWiener$Ly, sampWiener$Lt)
res

}
