% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.FSVD.R
\name{print.FSVD}
\alias{print.FSVD}
\title{Print an FSVD object}
\usage{
\method{print}{FSVD}(x, ...)
}
\arguments{
\item{x}{An FSVD object.}

\item{...}{Not used.}
}
\description{
Print a simple description of an FSVD object
}
\examples{
set.seed(1)
n <- 20
pts <- seq(0, 1, by=0.05)
sampWiener <- Wiener(n, pts)
sampWiener <- Sparsify(sampWiener, pts, 10)
res <- FSVD(sampWiener$Ly, sampWiener$Lt)
res

}

