% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.WFDA.R
\name{print.WFDA}
\alias{print.WFDA}
\title{Print a WFDA object}
\usage{
\method{print}{WFDA}(x, ...)
}
\arguments{
\item{x}{A WFDA object.}

\item{...}{Not used.}
}
\description{
Print a simple description of a WFDA object
}
