% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectK.R
\name{selectK}
\alias{selectK}
\title{Selects number of functional principal components for
given FPCA output and selection criteria}
\usage{
selectK(fpcaObj, criterion = "AIC", FVEthreshold = NULL, fixedK = NULL,
  y = NULL, t = NULL)
}
\arguments{
\item{fpcaObj}{A list containing FPCA related subjects returned by MakeFPCAResults().}

\item{criterion}{A string specifying selection criterion for number of functional principal components, available options: 'FVE', 'AIC', 'BIC', 'fixedK' - default: 'AIC'}

\item{FVEthreshold}{A threshold percentage specified by user when using "FVE" as selection criterion: (0,1] - default: NULL}

\item{fixedK}{An integer: user-specified number of components to be chosen - default: NULL}

\item{y}{A list of \emph{n} vectors containing the observed values for each individual - default: NULL}

\item{t}{A list of \emph{n} vectors containing the observation time points for each individual corresponding to y - default: NULL}
}
\value{
A list including the following two fields:
\item{k}{An integer indicating the selected number of components based on given criterion.}
\item{criterion}{The calculated criterion value for the selected number of components, i.e. FVE, AIC or BIC value, NULL for fixedK criterion.}
}
\description{
Selects number of functional principal components for
given FPCA output and selection criteria
}

