% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetOptions.R
\name{SetOptions}
\alias{SetOptions}
\title{Set the PCA option list}
\usage{
SetOptions(y, t, optns)
}
\arguments{
\item{y}{A list of \emph{n} vectors containing the observed values for each individual.}

\item{t}{A list of \emph{n} vectors containing the observation time points for each individual corresponding to y.}

\item{optns}{A list of options control parameters specified by \code{list(name=value)}. See `Details'.

See '?FPCA for more details. Casual users are not advised to tamper with this function.}
}
\description{
Set the PCA option list
}
