% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/sparsify.R
\name{sparsify}
\alias{sparsify}
\title{Sparsify densely observed functional data}
\usage{
sparsify(samp, pts, sparsity, aggressive = FALSE)
}
\arguments{
\item{samp}{A matrix of densely observed functional data, with each row containing one sample.}

\item{pts}{A vector of grid points corresponding to the columns of \code{samp}.}

\item{sparsity}{A vector of integers. The number of observation per sample is chosen to be one of the elements in sparsity with equal chance.}

\item{aggressive}{Sparsify in an "aggressive" manner making sure that near-by readings are excluded.}
}
\value{
A list of length 2, containing the following fields:
\item{tList}{A list of observation time points for each sample.}
\item{yList}{A list of values for each sample, corresponding to the time points.}
}
\description{
Given a matrix of densely observed functional data, make a sparsified sample.
}

