% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WFDA.R
\name{WFDA}
\alias{WFDA}
\title{Warped Functional DAta Analysis}
\usage{
WFDA(Ly, Lt, optns = list())
}
\arguments{
\item{Ly}{A list of \emph{n} vectors containing the observed values for each individual. Missing values specified by \code{NA}s are supported for dense case (\code{dataType='dense'}).}

\item{Lt}{A list of \emph{n} vectors containing the observation time points for each individual corresponding to y. Each vector should be sorted in ascending order.}

\item{optns}{A list of options control parameters specified by \code{list(name=value)}. See `Details'.}
}
\value{
A list containing the following fields: 
\item{lambda}{Penalty parameter used.}
\item{aligned}{Aligned curves evaluated at time 't'}
\item{h}{Warping functions for 't'} 
\item{hInv}{Inverse warping functions evaluated at 't'} 
\item{The mean cost associated with each curve}{costs} 
\item{timing}{The time required by time-warping.}
}
\description{
Pairwise curve synchronization for functional data
}
\examples{
N = 50;
eps = 0.05;
M = 51;
set.seed(123) 
me <- function(t) exp(-10*(t-0.5)^2);
T = seq(0,1,length.out = M) 
recondingTimesMat = matrix(nrow = N, ncol = M)
yMat = matrix(nrow = N, ncol = M)

for (i in 1:N){
  peak = runif(min = 0.3,max =  0.7,1)# rnorm(1, mean=0.5, sd= 0.133)
  recondingTimesMat[i,] = sort( c(0,    runif(min = 0, max = peak, n = round((M-3)/2)), 
                                  peak, runif(min = peak, max = 1, n = round((M-3)/2)), 1) )
  yMat[i,] = me(recondingTimesMat[i,])  + rnorm(M, mean=0.0, sd=  eps) 
}

Y = as.list(as.data.frame(t(yMat)))
X = rep(list(T),N)

sss =  WFDA(Y, X)
par(mfrow=c(1,2))
matplot(x= T, t(yMat), t='l'); grid()
matplot(x= T, t(sss$aligned), t='l'); grid() 
}
\references{
\cite{Tang, R. and Mueller, H.G. (2008). "Pairwise curve synchronization for functional data." Biometrika 95, 875-889}

\cite{Tang, R. and Mueller, H.G. (2009) "Time-synchronized clustering of gene expression trajectories." Biostatistics 10, 32-45}
}

