% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CheckOptions.R
\name{CheckOptions}
\alias{CheckOptions}
\title{Check option format}
\usage{
CheckOptions(t, optns, n)
}
\arguments{
\item{t}{is a n-by-1 list of vectors}

\item{optns}{is an initialized option list}

\item{n}{is a total number of sample curves}
}
\description{
Check if the options structure is valid and set the ones that are NULL
}
