% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateDiagnosticsPlot.R
\name{CreateDiagnosticsPlot}
\alias{CreateDiagnosticsPlot}
\title{Functional Principal Component Analysis Diagnostics plot}
\usage{
CreateDiagnosticsPlot(ret, openNewDev = FALSE)
}
\arguments{
\item{ret}{An FPCA class object returned by FPCA().}

\item{openNewDev}{A logical specifying if a new device should be opened - default: FALSE}
}
\description{
This function will open a new device if not instructed otherwise.
}
\examples{
set.seed(1)
n <- 20
pts <- seq(0, 1, by=0.05)
sampWiener <- Wiener(n, pts)
sampWiener <- Sparsify(sampWiener, pts, 10)
res <- FPCA(sampWiener$yList, sampWiener$tList, 
            list(dataType='Sparse', error=FALSE, kernel='epan', verbose=FALSE))
CreateDiagnosticsPlot(res)
}

