% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/createDiagnosticsPlot.R
\name{createDiagnosticsPlot}
\alias{createDiagnosticsPlot}
\title{Functional Principal Component Analysis Diagnostics plot}
\usage{
createDiagnosticsPlot(ret, openNewDev = TRUE)
}
\arguments{
\item{ret}{An FPCA class object returned by FPCA().}

\item{openNewDev}{A logical specifying if a new device should be opened - default: TRUE}

\item{t}{A list of \emph{n} vectors containing the observation time points for each individual corresponding to y.}
}
\description{
This function will open a new device if not instructed otherwise.
}
\examples{
set.seed(1)
n <- 20
pts <- seq(0, 1, by=0.05)
sampWiener <- wiener(n, pts)
sampWiener <- sparsify(sampWiener, pts, 10)
res <- FPCA(sampWiener$yList, sampWiener$tList, list(dataType='Sparse', error=FALSE, kernel='epan', verbose=TRUE))
createDiagnosticsPlot(sampWiener$tList, res)
}

