% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateFuncBoxPlot.R
\name{CreateFuncBoxPlot}
\alias{CreateFuncBoxPlot}
\title{Create functional boxplot using 'bagplot', 'KDE' or 'pointwise' methodology}
\usage{
CreateFuncBoxPlot(fpcaObj, addIndx = NULL, variant = "bagplot",
  unimodal = NULL, ifactor = NULL, k = length(fpcaObj$lambda), ...)
}
\arguments{
\item{fpcaObj}{A object of class FPCA returned by the function FPCA().}

\item{addIndx}{A vector of indeces corresponding to which samples one should overlay (Default: NULL)}

\item{variant}{A character variable indicating which methodology should be used ('kde', 'bagplot' or 'pointwise') to create the functional box-plot (Default: 'bagplot')}

\item{unimodal}{logical specifying if the KDE estimate should be unimodal (default: TRUE, relavant only for variant='KDE')}

\item{ifactor}{Inflation ifactor for the bag-plot defining the loop of bag-plot or multiplying ifactor the KDE pilot bandwidth matrix.  (see ?aplpack::compute.bagplot/?ks::Hpi; default: 2.58/2).}

\item{k}{The integer number of the first k components used for the representation. (default: length(fpcaObj$lambda ))}

\item{...}{Additional arguments for the 'plot' function.}
}
\description{
Using an FPCA object create a functional box-plot based on the function scores.
}
\examples{
set.seed(1)
n <- 20
pts <- seq(0, 1, by=0.05)
sampWiener <- Wiener(n, pts)
sampWiener <- Sparsify(sampWiener, pts, 10)
res <- FPCA(sampWiener$yList, sampWiener$tList, 
            list(dataType='Sparse', error=FALSE, kernel='epan', verbose=TRUE))
CreateFuncBoxPlot(res, addIndx=c(1:3)) 
}
\references{
\cite{P. J. Rousseeuw, I. Ruts, J. W. Tukey (1999): The bagplot: a bivariate boxplot, The American Statistician, vol. 53, no. 4, 382-387}
}

