% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createOutliersPlot.R
\name{createOutliersPlot}
\alias{createOutliersPlot}
\title{Functional Principal Component Scores Plot}
\usage{
createOutliersPlot(fpcaObj, optns, ...)
}
\arguments{
\item{fpcaObj}{An FPCA class object returned by FPCA().}

\item{optns}{A list of options control parameters specified by \code{list(name=value)}. See `Details'.}

\item{...}{Additional arguments for the 'plot' function.}
}
\description{
This function will create using the first two FPC scores a set of convex hulls of the scores as these hulls are defined by the references.
}
\details{
Available control options are 
\describe{
\item{ifactor}{inflation ifactor for the bag-plot defining the loop of bag-plot or multiplying ifactor the KDE pilot bandwidth matrix. (see ?aplpack::compute.bagplot; ?ks::Hpi respectively; default: 2.58; 2 respectively).}
\item{variant}{string defining the outlier method used ('KDE' or 'bagplot') (default: 'KDE')}
\item{unimodal}{logical specifying if the KDE estimate should be unimodal (default: FALSE, relavant only for variant='KDE')}
\item{nSlices}{integer between 3 and 16, denoting the number of slices to be used, default 4, relavant only for groupingType='slice' }
\item{groupingType}{string specifying if a slice grouping ('slice') or a standard percentile/bagplot grouping ('standard') should be returned (default: 'standard')} 
}
}
\examples{
set.seed(1)
n <- 20
pts <- seq(0, 1, by=0.05)
sampWiener <- wiener(n, pts)
sampWiener <- sparsify(sampWiener, pts, 10)
res <- FPCA(sampWiener$yList, sampWiener$tList, 
            list(dataType='Sparse', error=FALSE, kernel='epan', verbose=TRUE))
createOutliersPlot(res)
}
\references{
\cite{P. J. Rousseeuw, I. Ruts, J. W. Tukey (1999): The bagplot: a bivariate boxplot, The American Statistician, vol. 53, no. 4, 382-387}
\cite{R. J. Hyndman and H. L. Shang. (2010) Rainbow plots, bagplots, and boxplots for functional data, Journal of Computational and Graphical Statistics, 19(1), 29-45}
}

