% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FkC.R
\name{FkC}
\alias{FkC}
\title{Functional clustering and identifying substructures of longitudinal data}
\usage{
FkC(y, t, k = 3, maxIter = 20, optns = list(maxK = 3, lean = TRUE))
}
\arguments{
\item{y}{A list of \emph{n} vectors containing the observed values for each individual. Missing values specified by \code{NA}s are supported for dense case (\code{dataType='dense'}).}

\item{t}{A list of \emph{n} vectors containing the observation time points for each individual corresponding to y.}

\item{k}{A scalar defining the number of clusters to define; default 3.}

\item{maxIter}{A scalar defining the maximum number of iterations allowed; default 20, common for both the simple kmeans initially and the subsequent k-centres}

\item{optns}{A list of options control parameters specified by \code{list(name=value)}; by default: 'maxK' is set to 3. See `Details in ?FPCA'.}
}
\value{
A list containing the following fields:
\item{cluster}{A vector of levels 1:k, indicating the cluster to which each curve is allocated.} 
\item{fpcaList}{A list with the fpcaObj for each separate cluster.} 
\item{iterToConv}{A number indicating how many iterations where required until convergence.}
}
\description{
Functional clustering and identifying substructures of longitudinal data
}
\examples{
set.seed(1)
n <- 251
pts <- seq(0, 1, by=0.01)
sampWiener <- Wiener(n, pts)
sampWiener <- Sparsify(sampWiener, pts, 101) 
fkcObj <- FkC(sampWiener$Ly, sampWiener$Lt)
}
\references{
\cite{Jeng-Min Chiou, Pai-Ling Li, "Functional clustering and identifying substructures of longitudinal data." Journal of the Royal Statistical Society 69 (2007): 679-699}
}

