% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/FPCAder.R
\name{FPCAder}
\alias{FPCAder}
\title{Functional Principal Component Analysis Derivatives}
\usage{
FPCAder(fpcaObj, ...)
}
\arguments{
\item{fpcaObj}{A object of class FPCA returned by the function FPCA().}
}
\description{
Derivative FPCA for dense or sparse functional data.
}
\examples{
set.seed(1)
n <- 20
pts <- seq(0, 1, by=0.05)
sampWiener <- wiener(n, pts)
sampWiener <- sparsify(sampWiener, pts, 10)
res <- FPCA(sampWiener$yList, sampWiener$tList, list(dataType='Sparse', error=FALSE, kernel='epan', verbose=TRUE))
resder <- FPCAder(res)
}
\references{
\cite{Liu, Bitao, and Hans-Georg Müller. "Estimating derivatives for samples of sparsely observed functions, with application to online auction dynamics." Journal of the American Statistical Association 104, no. 486 (2009): 704-717. (Sparse data FPCA)}
}

