% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/FPCAregScalar.R
\name{FPCAregScalar}
\alias{FPCAregScalar}
\title{Functional Principal Component Analysis Regression with Scalar dependent variable}
\usage{
FPCAregScalar(fpcaObjList, extVar = NULL, depVar, varSelect = NULL,
  bootStrap = FALSE, regressionType = NULL, y = NULL, t = NULL,
  lambda = 1e-09, ...)
}
\arguments{
\item{extVar}{A data.frame holding external explanatory variables.}

\item{depVar}{A vector with the dependant variable.}

\item{varSelect}{A string defining the type of step-wise variable selection applied ('AIC' or 'BIC'); this calls 'MASS::stepAIC()'. (default: NULL)}

\item{regressionType}{A string defining the type of regression to perform ('dense' or 'sparse'); (default : automatically determined based on 'fpcaObj'}

\item{y}{A list of \emph{n} vectors containing the observed values for each individual.}

\item{t}{A list of \emph{n} vectors containing the observation time points for each individual corresponding to y.}

\item{lambda}{A scalar for the ridge correction in the sparse regression}

\item{...}{Additional arguments}

\item{fpcaObj}{A object of class FPCA returned by the function FPCA().}
}
\description{
Functional regression for dense or sparse functional data.
}
\references{
\cite{Yao, F., Mueller, H.G., Wang, J.L. "Functional Linear Regression Analysis for Longitudinal Data." Annals of Statistics 33, (2005): 2873-2903.(Dense data)}
\cite{Senturk, D., Nguyen, D.V. "Varying Coefficient Models for Sparse Noise-contaminated Longitudinal Data", Statistica Sinica 21(4), (2011): 1831-1856. (Sparse data)}
}

