% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getCrCorYX.R
\name{getCrCorYX}
\alias{getCrCorYX}
\title{Make cross-correlation matrix from cross-covariance matrix}
\usage{
getCrCorYX(ccXY, ccXX, ccYY)
}
\arguments{
\item{ccXY}{The cross-covariance matrix between variables X and Y.}

\item{ccXX}{The auto-covariance matrix of variable X or the diagonal of that matrix.}

\item{ccYY}{The auto-covariance matrix of variable Y or the diagonal of that matrix.}
}
\description{
Make cross-correlation matrix from cross-covariance matrix
}

