% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNormalisedSample.R
\name{getNormalisedSample}
\alias{getNormalisedSample}
\title{Normalise sparse functional sample}
\usage{
getNormalisedSample(fpcaObj)
}
\arguments{
\item{fpcaObj}{An FPCA object.}
}
\value{
A list containing the normalised sample 'y' at times 't'
}
\description{
Normalise sparse functional sample given in an FPCA object
}

