% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetNormalisedSample.R
\name{GetNormalisedSample}
\alias{GetNormalisedSample}
\title{Normalise sparse functional sample}
\usage{
GetNormalisedSample(fpcaObj)
}
\arguments{
\item{x}{An FPCA object.}

\item{...}{Not used.}
}
\description{
Normalise sparse functional sample given in an FPCA object
}

