% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetNormalisedSample.R
\name{GetNormalisedSample}
\alias{GetNormalisedSample}
\title{Normalise sparse functional sample}
\usage{
GetNormalisedSample(fpcaObj)
}
\arguments{
\item{fpcaObj}{An FPCA object.}
}
\value{
A list containing the normalised sample 'y' at times 't'
}
\description{
Normalise sparse functional sample given in an FPCA object
}

