% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HandleNumericsAndNAN.R
\name{HandleNumericsAndNAN}
\alias{HandleNumericsAndNAN}
\title{Check if NaN are present in the data and if yes remove them}
\usage{
HandleNumericsAndNAN(y, t)
}
\arguments{
\item{y}{is a n-by-1 list of vectors}

\item{t}{is a n-by-1 list of vectors}
}
\description{
Check if there are problems cause by missing values with the form and basic structure of the functional data 'y' and the recorded times 't'.
}

