% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormCurveToArea.R
\name{NormCurvToArea}
\alias{NormCurvToArea}
\title{Normalise a curve to a particular area, by multiplication with a factor}
\usage{
NormCurvToArea(y, x = seq(0, 1, length.out = length(y)), area = 1)
}
\arguments{
\item{y}{values of curve at time-points x}

\item{x}{design time-points (default: seq(0,1, length.out=length(y)))}

\item{area}{value to normalise the curve onto (default: 1)}
}
\value{
yNew values of curve at times x such that [\\int{yNew}dx = area]
}
\description{
Normalise a curve such that \\int{yNew}dx = area (according to trapezoid integration)
}
