% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.FPCA.R
\name{print.FPCA}
\alias{print.FPCA}
\title{Print an FPCA object}
\usage{
print.FPCA(x, ...)
}
\arguments{
\item{x}{An FPCA object.}

\item{...}{Not used.}
}
\description{
Print a simple description of an FPCA object
}
\examples{
set.seed(1)
n <- 20
pts <- seq(0, 1, by=0.05)
sampWiener <- wiener(n, pts)
sampWiener <- sparsify(sampWiener, pts, 10)
res <- FPCA(sampWiener$yList, sampWiener$tList, 
            list(dataType='Sparse', error=FALSE, kernel='epan', verbose=TRUE))
res

}

