% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/print.FPCA.R
\name{print.FPCA}
\alias{print.FPCA}
\title{Print an FPCA object}
\usage{
print.FPCA(x, ...)
}
\arguments{
\item{x}{An FPCA object.}

\item{...}{Not used.}
}
\description{
Print a simple description of an FPCA object
}

